Use fabs() instead of abs().
2005-09-08  Behdad Esfahbod  <behdad@gnome.org>
        * Src/DasherCore/DasherViewSquare.inl: Use fabs() instead of abs().

        * Src/DasherCore/Parameters.h: Use const char * instead of char * in
        structs.

        * Src/DasherCore/UserLocation.cpp: Cast integer printf arguments
        for %0.4f to double.

+++ Src/DasherCore/UserLocation.cpp
@@ -242,7 +242,7 @@ string CUserLocation::GetTabMouseXY(bool bReturnNormalized)
   if (bReturnNormalized)
     sprintf(szNum, "%0.4f\t%0.4f\n", m_dNormalizedLocationX, m_dNormalizedLocationY);
   else
-    sprintf(szNum, "%0.4f\t%0.4f\n", m_iLocationX, m_iLocationY);
+    sprintf(szNum, "%0.4f\t%0.4f\n", (double)m_iLocationX, (double)m_iLocationY);
 
   strResult += szNum;
